PRE-RELEASE 0.2
This is the first eggcode version. email me at elijah.plays34@gmail.com if there are any issues. (there probably will be)
Eggcode is a basic, bare bones coding language made in python for beginners. It works. I try to update it as much as I can. It runs variables set by the user and some other cool stuff. It has a "batch" command that works like Commodore Basic. (or for newer coders, .bat files) Run the .py file in a python interpreter.